dtm gefixt

+++ src/core/DeltaTimeMachine.h
@@ -12,7 +12,7 @@ namespace DT {
                                    //do not call the start function inside the game loop as this will destroy your reference point
                                    //then at the end of your game loop you call the update function
                                    // if at any point in the game loop you need the delta time variable call the getdt function which will give you a float
-        float dt=1;
+        double dt=1;
         typedef std::chrono::time_point<std::chrono::system_clock,std::chrono::duration<long long,std::ratio<1,1000000000>>> ctime;
         ctime frame1;
         ctime frame2;
